// src/types/api/api-about-us-data.ts
export interface Root {
  data: Data;
  meta: Meta;
}

export interface Data {
  id: number;
  attributes: Attributes;
}

export interface Attributes {
  createdAt: string;
  updatedAt: string;
  publishedAt: string;
  locale: string;
  section_1: Section1;
  block_1: Block1;
  block_2: Block2;
  block_with_background: BlockWithBackground;
  images: Images;
  localizations: Localizations;
}
// ...
export interface Meta {}