// src/types/api/api-premium-service-data.ts
export interface Root {
  data: Data;
  meta: Meta;
}

export interface Data {
  id: number;
  attributes: Attributes;
}

export interface Attributes {
  createdAt: string;
  updatedAt: string;
  publishedAt: string;
  locale: string;
  hero: Hero;
  block_1: Block1;
  block_2: Block2;
  localizations: Localizations;
}
// ...
export interface Meta {}