// src/types/api/api-destino-page-data.ts
export interface Root {
  data: Data;
  meta: Meta;
}

export interface Data {
  id: number;
  attributes: Attributes;
}

export interface Attributes {
  title: string;
  text: string;
  createdAt: string;
  updatedAt: string;
  publishedAt: string;
  locale: string;
  miami_text: string;
  tulum_text: string;
  ibiza_text: string;
  pde_text: string;
  images_ibiza: ImagesIbiza;
  images_tulum: ImagesTulum;
  images_punta_del_este: ImagesPuntaDelEste;
  images_miami: ImagesMiami;
  clickable_services: ClickableServices;
  localizations: Localizations2;
}
// ...
export interface Meta {}